Refactor PUSH to use inline i16 encoding with auto-promotion to PUSHC
PUSH now encodes a signed 16-bit immediate (3 bytes: opcode + 2-byte BE
i16) instead of a 64-bit one (9 bytes). Values outside the i16 range
(-32768..=32767) automatically promote to PUSHC via the constant pool.

InstructionBuilder::push(imm: i64) dispatches: small values emit
Push { imm: i16 }, large values call push_const(imm) -> PushC.

Assembler special-cases the PUSH mnemonic so that PUSH 100000 promotes
to PUSHC rather than failing with IntegerOutOfRange.

Disassembler special-cases Instruction::Push to bypass the i16
FmtOperand label-resolution logic and render the immediate as plain
decimal.

All affected byte-offset constants in stream, codec, asm, and disasm
tests are updated to reflect the 9->3 byte size reduction.

+++ crates/bytecode/src/types/instruction.rs
@@ -176,10 +176,10 @@ mod tests {
 
     #[test]
     fn push_imm_field() {
-        let p = Instruction::Push { imm: i64::MIN };
+        let p = Instruction::Push { imm: i16::MIN };
         assert_eq!(p.opcode(), Opcode::Push);
         if let Instruction::Push { imm } = p {
-            assert_eq!(imm, i64::MIN);
+            assert_eq!(imm, i16::MIN);
         } else {
             panic!("pattern match failed");
         }
